+++ src/mocks/offersData.js
@@ -236,39 +236,5 @@ export default [
     "rating": 4.5,
     "title": `Beautiful & luxurious studio at great location`,
     "type": `apartment`
-  },
-  {
-    "bedrooms": 1,
-    "city": {
-      "location": {
-        "latitude": 52.3909553943508,
-        "longitude": 4.85309666406198,
-        "zoom": 12
-      },
-      "name": `Paris`
-    },
-    "description": `A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.`,
-    "goods": [`Heating`, `Kitchen`, `Cable TV`, `Washing machine`, `Coffee machine`, `Dishwasher`],
-    "host": {
-      "avatar_url": `./img/avatar-angelina.jpg`,
-      "id": 3,
-      "is_pro": true,
-      "name": `Angelina`
-    },
-    "id": 8,
-    "images": [`./img/room.jpg`, `./img/apartment-01.jpg`, `./img/apartment-02.jpg`, `./img/apartment-03.jpg`, `./img/apartment-02.jpg`, `./img/apartment-03.jpg`],
-    "is_favorite": true,
-    "is_premium": false,
-    "location": {
-      "latitude": 52.3809553943508,
-      "longitude": 4.939309666406198,
-      "zoom": 8
-    },
-    "max_adults": 3,
-    "preview_image": `./img/apartment-01.jpg`,
-    "price": 90,
-    "rating": 3.5,
-    "title": `Beautiful & luxurious studio at great location`,
-    "type": `apartment`
   }
 ];
