+++ src/utils/common.js
@@ -1,7 +1,5 @@
 import offersData from '../mocks/offersData';
 
-import {DEFAULT_CURRENT_CITY} from "../const";
-
 export const getRatingStarValue = (number) => {
   return Math.round(number) * 20 + `%`;
 };
@@ -45,27 +43,6 @@ export const adaptCommentsData = (data) => ({
 
 export const getOffers = () => offersData.map(adaptOfferData); // временно. пока нет api
 
-export const getOffersByCity = (city) => {
-  return getOffers().filter((offer) => offer.city.name === city.name);
-};
-
-export const getCurrentCity = (state) => {
-  if (state && state.CITY.city) {
-    return state.CITY.city;
-  }
-  return getCities()
-    .find((city) => city.name === DEFAULT_CURRENT_CITY);
-};
-
-export const getCities = () => {
-  const cities = getOffers()
-    .reduce((generalOffer, offer) => {
-      if (!generalOffer.hasOwnProperty(offer.city.name)) {
-        generalOffer[offer.city.name] = {...offer.city};
-      }
-
-      return generalOffer;
-    }, {});
-
-  return Object.values(cities);
+export const getOffersByCity = (state) => {
+  return getOffers().filter((offer) => offer.city.name === state.CITY.name);
 };
